day 6

+++ respectablefilename.py
@@ -18,6 +18,7 @@ class star():
     plane = None
     lvl = 0
 
+    
     def forlatersql(self):
         connection = sqlite3.connect('dbase.db')
         cursor = connection.cursor()
@@ -28,14 +29,11 @@ class star():
         query = "select * from stars"
         cursor.execute(query)
 
-
-        
-
     def __init__(self):
 
         
         def randname():
-                letters = ["bdfghjklmnprstvwyz","aeiou"]
+                letters = ["bdfgjklmnprstvwyz","aeiou"]
                 name = ''.join(random.choice(letters[random.randint(0,1)]) for _ in range(random.randint(3,8))).capitalize()
 
                 return name
@@ -57,7 +55,7 @@ class star():
         if self.govt < 0: self.govt = 0
         self.lawlvl = (random.randint(1,6) + random.randint(1,6) - 7 + self.govt)
         if self.atm <= 0: self.atm = 0
-        
+        if self.lawlvl < 0: self.lawlvl = 0
         
         line = open("tables.csv","r").read().split("\n")
             
@@ -80,11 +78,9 @@ class star():
                 if tempstar == "C" or tempstar == "D" or tempstar == "E"or tempstar == "X":
                     return False
 
-
             info = json.loads(line[twodie-2+multi])
             return info[varname]
-
-        
+    
         def techlvl():
             tech = random.randint(1,6)
             data = open("table2.csv","r").read().split("\n")
@@ -100,9 +96,7 @@ class star():
                 k = json.loads(data[nums[count]])
                 tech +=k[i]
                 count+= 1
-            return tech
-        
-
+            return tech 
         self.starport = info('starport')
         self.naval = info('naval')
         self.scout = info('scout')
@@ -110,17 +104,20 @@ class star():
         self.plane = info('plane')
         self.lvl = techlvl()
         if self.lvl < 0: self.lvl
-        
-        self.forlatersql()
-
-        
-def main():   
+        self.forlatersql()     
+bignum = 0
+numba = 10        
+def main(): 
+    global bignum
+    global numba
     data = ["Name","Size","Atmosphere","Hydration","Population","Government","Law","Technology","Starport","Naval Base exists","Scout exists","Gas giant exists","Planetoids exist"]   
     connection = sqlite3.connect('dbase.db')
     cursor = connection.cursor()
-    cursor.execute('drop table stars')
-    connection.commit()
-    
+    if bignum == 0:       
+        cursor.execute('drop table stars')
+        connection.commit()
+    else:pass   
+    bignum+=1
     query = """
         create table if not exists stars (
             id integer primary key autoincrement,
@@ -138,46 +135,57 @@ def main():
             gas bool,
             plane bool);"""
     cursor.execute(query)
-    
-
-    for i in range(10):
-        star()
-    
-    connection = sqlite3.connect('dbase.db')
-    cursor = connection.cursor()
-    query = "select * from stars"
-    cursor.execute(query)
-    res = cursor.fetchall()
-    for i in res:
-        print(i)
+    for i in range(numba):star()
     while True:
-        inp = input("what do you want to look for?: ").strip()
-        inp2 = input("Enter value: ").strip()
-        try:
-            query = f"select * from stars where {inp} = '{inp2}'"
-            cursor.execute(query)
-            result = cursor.fetchall()
-            if len(result) == 0:
-                print("No data with that value")
-                continue
-            break
-        except:
-            print("invalid input")
-            print("Valid inputs: name, size, atm, hyd, pop, govt, law, tech, starport, naval, scout, gas, plane")
+        connection = sqlite3.connect('dbase.db')
+        cursor = connection.cursor()
+        query = "select * from stars"
+        cursor.execute(query)
+        result = cursor.fetchall()
+        
+        for i in (result):
+            for k in data:
+                temp = data.index(k)+1
+                if temp%2 == 0:
+                    print(f"\033[1m{k}\033[0m: {i[data.index(k)+1]}")
+                else: print(f"\033[1m{k:<20}\033[0m: {i[data.index(k)+1]:<20}",end="")
+            print("\n")
+    
+        choice = input('\n\033[1mWhat would you like to do? (help for options): \033[0m ')
+        if choice == 'help':print("\nYou can: \nlook: To look at star systems\ngenerate: To create new star systems")
+        
+        elif choice == 'look':
+            inp = input("\n\033[1mWould you like to look for name, size, atm, hyd, pop, \ngovt, law, tech, starport, naval, scout, gas, plane or all?: \033[0m").strip()
+            if inp == 'all':
+                inp2 = None
+                pass
+                
+            else: inp2 = input("\n\033[1mEnter value:\033[0m ").strip()
+            try:
+                
+                query = f"select * from stars where {inp} = '{inp2}'"
+                if inp == 'all': query = "select * from stars"
+                cursor.execute(query)
+                result = cursor.fetchall()
+                if len(result) == 0:
+                    print("\033[1mNo data with that value\033[0m ")
+                    continue
+                
+            except:
+                print("\033[1minvalid input\033[0m ")
 
-    
-    
-    
-    for i in result:
-        for k in data:
-            print(f"\033[1m{k}\033[0m: {i[data.index(k)+1]}")
-        print("\n")
-    
-    
+            for i in (result):
+                
+                for k in data:
+                    temp = data.index(k)+1
+                    if temp%2 == 0:
+                        print(f"\033[1m{k}\033[0m: {i[data.index(k)+1]}")
+                    else: print(f"\033[1m{k:<20}\033[0m: {i[data.index(k)+1]:<20}",end="")
+                print("\n")
+        
+        elif choice == 'generate':
+            numba = int(input("\033[1mHow many stars would you like to create?: \033[0m "))
+            for i in range(numba):
+                star()
 
 main()
-
-
-
-
-
